+++ cmd/native-relayer/cli/utils.go
@@ -3,31 +3,9 @@ package cli
 import (
 	"flag"
 	"fmt"
-	"time"
 
 	"github.com/spf13/viper"
 )
-
-// Configuration struct
-type Config struct {
-	NativeRPC                string        `mapstructure:"native_rpc"`
-	NativeGRPC               string        `mapstructure:"native_grpc"`
-	StartBlockHeight         int           `mapstructure:"start_block_height"`
-	IkaRPC                   string        `mapstructure:"ika_rpc"`
-	IkaSignerMnemonic        string        `mapstructure:"ika_signer_mnemonic"`
-	IkaNativeLcPackage       string        `mapstructure:"ika_native_lc_package"`
-	IkaNativeLcModule        string        `mapstructure:"ika_native_lc_module"`
-	IkaNativeLcFunction      string        `mapstructure:"ika_native_lc_function"`
-	IkaGasAcc                string        `mapstructure:"ika_gas_acc"`
-	IkaGasBudget             string        `mapstructure:"ika_gas_budget"`
-	BtcRpcHost               string        `mapstructure:"btc_rpc_host"`
-	BtcRpcUser               string        `mapstructure:"btc_rpc_user"`
-	BtcRpcPass               string        `mapstructure:"btc_rpc_pass"`
-	BtcConfirmationThreshold uint8         `mapstructure:"btc_confirmation_threshold"`
-	ProcessTxsInterval       time.Duration `mapstructure:"process_txs_interval"`
-	ConfirmTxsInterval       time.Duration `mapstructure:"confirm_txs_interval"`
-	DbFile                   string        `mapstructure:"db_file"`
-}
 
 func loadConfig() (*Config, error) {
 	var configFile string
@@ -44,8 +22,7 @@ func loadConfig() (*Config, error) {
 	}
 
 	if err := viper.ReadInConfig(); err != nil {
-		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
-		} else {
+		if _, ok := err.(viper.ConfigFileNotFoundError); !ok {
 			return nil, fmt.Errorf("error reading config file: %w", err)
 		}
 	}
